expert-guided regressions and clustering
However, it is such a small fraction of sites it does not actually help much,
and I feel it adds more compelxity to the explanation than is worht it for
communication.

However, one thing we might wat to to think about is doing some type of
clustered regression with the slopes of variables. Perhaps this
could be done with stan or infer.NET, but it would be quite a bigger project