alternately cluster based on normalized data rather than ranked.
this would be more relatable to the community, and I'm also struggling
to articulare my reasons why the ranking approach feels right to me.

I was also surprised: there si very little difference between the two,
which si good cause we odn't want the method to be desntivie. the one
exceptions i elko, which is noticeably worse for the normalized approahc.

Still think I should use it though, even with that degradataion; that cost
is worth the communication benefit.